去掉PCH.h，使用Config.h。重构Foundation内的Config和Log不依赖平台实现

+++ src/render/renderer/render_device.cpp
@@ -440,7 +440,7 @@ void RenderDevice::DrawRenderOBject(render::RenderObject* ro, render::ShaderAsse
 	mGraphicCmd->SetIndexBuffer(ib);
 	mGraphicCmd->SetDrawPrimitiveTopology(
 		RHIPrimitiveTopology::TriangleList);
-	mGraphicCmd->DrawIndexedInstanced(indexCount, 1, 0, 0, 0);
+	mGraphicCmd->DrawIndexedInstanced((uint32_t)indexCount, 1, 0, 0, 0);
 
 }
